BORKED: cute new process output
I tried making cute process output, but its out of order, thats because
the loop executes and queues up a bunch of requests and then the "then"
block fires. but what we really want is to do is wait till all of those
finish before executing the "then" block.

I also think that the requests are being executed in the "wrong" order.
Might be able to solve this with a simple timeout 1000s. But could be
more sophisticated and do an actual request queue?